Moving autograd from header only lib to a compiled lib
