+++ include/af/autograd/Functions.hpp
@@ -8,31 +8,12 @@
  ********************************************************/
 #pragma once
 
-#include <af/autograd/Variable.hpp>
-
 namespace af {
     namespace autograd {
 
-        Variable operator +(const Variable lhs, const Variable rhs)
-        {
-            auto result = lhs.array() + rhs.array();
-            auto grad_func = [](std::vector<Variable> inputs, Variable grad_output) {
-                inputs[0].addGrad(grad_output);
-                inputs[1].addGrad(grad_output);
-            };
-            return Variable(result, {lhs, rhs}, grad_func);
-        }
+        class Variable;
 
-        Variable operator *(const Variable lhs, const Variable rhs)
-        {
-            auto result = lhs.array() * rhs.array();
-            auto grad_func = [](std::vector<Variable> inputs, Variable grad_output) {
-                inputs[0].addGrad(grad_output * inputs[1]);
-                inputs[1].addGrad(grad_output * inputs[0]);
-            };
-            return Variable(result, {lhs, rhs}, grad_func);
-        }
-
+        Variable operator +(const Variable lhs, const Variable rhs);
+        Variable operator *(const Variable lhs, const Variable rhs);
     }
-    namespace ag = autograd;
 }
