+++ src/autograd/Functions.cpp
@@ -1,0 +1,37 @@
+/*******************************************************
+ * Copyright (c) 2017, ArrayFire
+ * All rights reserved.
+ *
+ * This file is distributed under 3-clause BSD license.
+ * The complete license agreement can be obtained at:
+ * http://arrayfire.com/licenses/BSD-3-Clause
+ ********************************************************/
+
+#include <af/autograd/Variable.hpp>
+#include <af/autograd/Functions.hpp>
+
+namespace af {
+    namespace autograd {
+
+        Variable operator +(const Variable lhs, const Variable rhs)
+        {
+            auto result = lhs.array() + rhs.array();
+            auto grad_func = [](std::vector<Variable> inputs, Variable grad_output) {
+                inputs[0].addGrad(grad_output);
+                inputs[1].addGrad(grad_output);
+            };
+            return Variable(result, {lhs, rhs}, grad_func);
+        }
+
+        Variable operator *(const Variable lhs, const Variable rhs)
+        {
+            auto result = lhs.array() * rhs.array();
+            auto grad_func = [](std::vector<Variable> inputs, Variable grad_output) {
+                inputs[0].addGrad(grad_output * inputs[1]);
+                inputs[1].addGrad(grad_output * inputs[0]);
+            };
+            return Variable(result, {lhs, rhs}, grad_func);
+        }
+
+    }
+}
